aminFormAndDebug
