+++ services.html
@@ -161,11 +161,11 @@
                   <option value="service">Есть</option>
                   <option value="category">Нет</option>
               </select>-->
-              <div>
+              <div class="parent-existence">
                 <input class="add-extra" type="radio" id="type" name="drone" value="service" name="type" required/>
                 <label for="type">Есть</label>
               </div>
-              <div>
+              <div class="parent-existence">
                 <input class="hide-extra" type="radio" id="type" name="drone" value="category" name="type" required/>
                 <label for="type">Нет</label>
               </div>
